+++ src/mbart/mod.rs
@@ -1,4 +1,5 @@
 mod attention;
+mod decoder;
 mod embeddings;
 mod encoder;
 mod mbart_model;
@@ -7,4 +8,6 @@ pub use mbart_model::{
     MBartConfig, MBartConfigResources, MBartModelResources, MBartVocabResources,
 };
 
+pub use attention::LayerState;
+pub(crate) use decoder::MBartDecoderLayer;
 pub(crate) use encoder::MBartEncoderLayer;
